Fixed audio paths, updated readme.
