+++ generic_parser.py
@@ -104,6 +104,8 @@ class TableParser:
         """
         table_basename = os.path.basename(table_path)
         table_subpath = os.path.dirname(table_path)[len(tables_base_path):]
+        while table_subpath.startswith(os.sep):
+            table_subpath = table_subpath[1:]
         audio_rel_no_ext_paths = os.path.join(table_subpath, table_basename.split(".")[0])
         with open(table_path) as fp:
             csvr = csv.reader(fp, delimiter=self.delimiter)
